Code climate - try to reduce duplication
Also fix warning on CardMetaData test
