+++ src/Components/Shared/CardMetaData/index.js
@@ -3,19 +3,22 @@ import PropTypes from 'prop-types'
 import MetaDataField from 'Components/Shared/MetaDataField'
 
 const CardMetaData = ({ metadata }) => {
-  return (
-    <dl className='details'>
-      {
-        metadata.map((md, index) => {
-          return <MetaDataField metadata={md} key={index} />
-        })
-      }
-    </dl>
-  )
+  if (metadata) {
+    return (
+      <dl className='details'>
+        {
+          metadata.map((md, index) => {
+            return <MetaDataField metadata={md} key={index} />
+          })
+        }
+      </dl>
+    )
+  }
+  return null
 }
 
 CardMetaData.propTypes = {
-  metadata: PropTypes.array.isRequired,
+  metadata: PropTypes.array,
 }
 
 export default CardMetaData
